Relocate ServerSelectionRule enum to config package
